+++ config/webpack/loaders/devSiteExample.js
@@ -12,7 +12,6 @@ const loader = async function loader() {
   const exampleSource = this.resourcePath;
   const parsedResourcePath = path.parse(exampleSource);
 
-  let code;
   let cssFileName;
   let exampleCssSource;
 
@@ -38,19 +37,28 @@ const loader = async function loader() {
     console.error(err);
   }
 
-  // The following code is a wip, it can probably be refactored for efficiency
-  if (cssFileName !== undefined) {
-    exampleCssSource = path.join(parsedResourcePath.dir, cssFileName);
-  }
-
-  code = [
+  const code = [
     'import React from \'react\';',
     `import Example from '${exampleSource}';`,
-    `import Css from '${exampleCssSource}?dev-site-codeblock';`,
     `import Code from '${exampleSource}?dev-site-codeblock';`,
     'import ExampleTemplate from \'terra-dev-site/lib/loader-components/_ExampleTemplate\';',
     '',
-    `export default ({ title, description, isCssExpanded, isExpanded }) => (
+    `export default ({ title, description, isExpanded }) => (
+      <ExampleTemplate
+        title={ title || '${startCase(parsedResourcePath.name)}'}
+        description={description}
+        example={<Example />}
+        exampleSrc={<Code />}
+        isExpanded={isExpanded}
+      />
+    );`,
+  ];
+
+  if (cssFileName !== undefined) {
+    exampleCssSource = path.join(parsedResourcePath.dir, cssFileName);
+    code.splice(2, 0, `import Css from '${exampleCssSource}?dev-site-codeblock';`);
+    code.pop();
+    code.push(`export default ({ title, description, isCssExpanded, isExpanded }) => (
       <ExampleTemplate
         title={ title || '${startCase(parsedResourcePath.name)}'}
         description={description}
@@ -60,15 +68,10 @@ const loader = async function loader() {
         isCssExpanded={isCssExpanded}
         isExpanded={isExpanded}
       />
-    );`,
-  ].join('\n');
-
-  if (exampleCssSource === undefined) {
-    code = code.replace('import Css from \'undefined?dev-site-codeblock\';', '');
-    code = code.replace('exampleCssSrc={<Css />}', '');
+    );`);
   }
 
-  return callback(null, code);
+  return callback(null, code.join('\n'));
 };
 
 module.exports = loader;