+++ src/navigation/_CollapsingNavigationMenu.jsx
@@ -1,3 +1,4 @@
+/* eslint-disable jsx-a11y/role-supports-aria-props */
 import React, {
   useState, useRef, useEffect, useLayoutEffect, useContext,
 } from 'react';